requirements.txt
fastapi
uvicorn
httpx
discord.py

